refactor: Move `OptionalValue` to shared `api-model`
This makes it available for components.
