+++ api/v1/model/build.gradle.kts
@@ -37,6 +37,8 @@ kotlin {
     sourceSets {
         commonMain {
             dependencies {
+                api(projects.shared.apiModel)
+
                 api(libs.kotlinxDatetime)
                 api(libs.konform)
