+++ core/src/main/kotlin/plugins/OpenApi.kt
@@ -45,9 +45,9 @@ import io.ktor.server.routing.routing
 import kotlinx.datetime.Instant
 
 import org.eclipse.apoapsis.ortserver.api.v1.model.CredentialsType
-import org.eclipse.apoapsis.ortserver.api.v1.model.OptionalValue
 import org.eclipse.apoapsis.ortserver.api.v1.model.RepositoryType
 import org.eclipse.apoapsis.ortserver.components.authorization.SecurityConfigurations
+import org.eclipse.apoapsis.ortserver.shared.apimodel.OptionalValue
 import org.eclipse.apoapsis.ortserver.utils.system.ORT_SERVER_VERSION
 
 import org.koin.ktor.ext.inject
